Minor fixes; add an endpoint POST /api/tweets/{tweet_id}/likes

+++ src/api/tweets_router.py
@@ -3,11 +3,14 @@ from typing import List, Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Response, status
 from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.exc import IntegrityError
 
 from src.database import queries as q
+from src.database import models
 from src.schemas import schemas
 from src.service.images import delete_images_by_ids, validate_images_in_db
-from src.service.web import check_api_key, check_tweet_id, get_session
+from src.service.web import check_api_key, check_tweet_exists, get_session
+from src.service.exceptions import ForbiddenError
 
 tweets_router: APIRouter = APIRouter()
 
@@ -24,7 +27,7 @@ logger = getLogger("routes_logger.tweets_logger")
                 "application/json": {
                     "example": {
                         "result": False,
-                        "error_type": "AuthenticationFailed",
+                        "error_type": "IdentificationError",
                         "error_message": "api_key <api_key> not exists",
                     }
                 }
@@ -36,7 +39,7 @@ logger = getLogger("routes_logger.tweets_logger")
                 "application/json": {
                     "example": {
                         "result": False,
-                        "error_type": "ValueError",
+                        "error_type": "HTTPException",
                         "error_message": "error message",
                     }
                 }
@@ -53,7 +56,7 @@ logger = getLogger("routes_logger.tweets_logger")
 async def create_new_tweet(
     tweet: schemas.Tweet,
     response: Response,
-    user_id: int = Depends(check_api_key),
+    user: models.User = Depends(check_api_key),
     session: AsyncSession = Depends(get_session),
 ):
     """
@@ -69,7 +72,7 @@ async def create_new_tweet(
             logger.debug("Some image ids not exists or relate other tweets")
             raise HTTPException(detail=str(exc), status_code=403)
 
-    tweet_id: int = await q.create_tweet(session, user_id, tweet.model_dump())
+    tweet_id: int = await q.create_tweet(session, user.id, tweet.model_dump())
     logger.debug("Tweet was created, tweet_id=%d", tweet_id)
     response.status_code = status.HTTP_201_CREATED
     return {"result": "true", "tweet_id": tweet_id}
@@ -85,7 +88,7 @@ async def create_new_tweet(
                 "application/json": {
                     "example": {
                         "result": False,
-                        "error_type": "AuthenticationFailed",
+                        "error_type": "IdentificationError",
                         "error_message": "api_key <api_key> not exists",
                     }
                 }
@@ -110,7 +113,7 @@ async def create_new_tweet(
                 "application/json": {
                     "example": {
                         "result": False,
-                        "error_type": "NotFound",
+                        "error_type": "NotFoundError",
                         "error_message": "tweet_id {tweet_id} not found",
                     }
                 }
@@ -123,8 +126,8 @@ async def create_new_tweet(
     },
 )
 async def delete_tweet(
-    tweet_id: int,
-    user_id: int = Depends(check_api_key),
+    tweet: models.Tweet = Depends(check_tweet_exists),
+    user: models.User = Depends(check_api_key),
     session: AsyncSession = Depends(get_session),
 ):
     """
@@ -132,22 +135,88 @@ async def delete_tweet(
     """
     logger.info("Start deleting the tweet")
 
-    # check tweet_id
-    await check_tweet_id(tweet_id, user_id, session)
-    logger.debug("tweet_id was found")
+    # check that tweet relates user
+    if tweet.user_id != user.id:
+        raise ForbiddenError(f"The tweet {tweet.user_id} does not belong to user {user.id}")
 
     # get images ids which relate this tweet
     images_ids: Optional[List[int]] = await q.get_images_ids_by_tweet_id(
-        session, tweet_id
+        session, tweet.id
     )
     logger.debug("List of images ids: %s", str(images_ids))
     # delete tweet and images from db
-    await q.delete_tweet_by_id(session, tweet_id)
+    await q.delete_tweet_by_id(session, tweet.id)
     logger.debug("Tweet and images were deleted from db")
 
     # delete images from disk
     if images_ids:
         await delete_images_by_ids(images_ids)
         logger.debug("Images were deleted from disk")
+
+    return {"result": True}
+
+
+@tweets_router.post(
+    "/api/tweets/{tweet_id}/likes",
+    status_code=200,
+    responses={
+        400: {
+            "description": "The tweet already has a user like",
+            "content": {
+                "application/json": {
+                    "example": {
+                        "result": False,
+                        "error_type": "HTTPException",
+                        "error_message": "The tweet {tweet.id} already has a user {user.id} like",
+                    }
+                }
+            },
+        },
+        401: {
+            "description": "api_key not exists",
+            "content": {
+                "application/json": {
+                    "example": {
+                        "result": False,
+                        "error_type": "IdentificationError",
+                        "error_message": "api_key {api_key} not exists",
+                    }
+                }
+            },
+        },
+        404: {
+            "description": "Tweet not found",
+            "content": {
+                "application/json": {
+                    "example": {
+                        "result": False,
+                        "error_type": "NotFoundError",
+                        "error_message": "tweet_id {tweet_id} not found",
+                    }
+                }
+            },
+        },
+        200: {
+            "description": "The user liked the tweet",
+            "content": {"application/json": {"example": {"result": True}}},
+        },
+    },
+)
+async def like_tweet(
+    tweet: models.Tweet = Depends(check_tweet_exists),
+    user: models.User = Depends(check_api_key),
+    session: AsyncSession = Depends(get_session),
+):
+    """
+    The endpoint deletes the tweet by id
+    """
+    logger.info("Start liking the tweet")
+    try:
+        await q.like_tweet(session, tweet, user)
+    except ValueError as exc:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=str(exc)
+        )
 
     return {"result": True}
